requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
rich>=13.0,<14.0
pytest>=7.0,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.0,<5.0
